02-langChain/2-Langchain Basics/2.4-VectorDatabases/requirements.txt
faiss-cpu

pydantic
python-dotenv
pypdf
bs4

langchain
langchain_core
langchain-community
langchain_huggingface
langchain_google_genai
langchain-pinecone
langchain-openai
langchain_groq

# google modules
langchain_google_genai
google-generativeai 

faiss-cpu
pymupdf
wikipedia
arxiv
defusedxml 

youtube_search
langchain-tavily
langgraph

chromadb
langchain_chromadb